requirements.txt
flask
Flask-SQLAlchemy
flask-login
pytest
coverage
